pyproject.toml
[tool.poetry]
name = "grocerystore"
version = "0.1.0"
description = ""
authors = ["Pampam000 <shilovmaxx.x@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
pillow = "^9.4.0"
django-debug-toolbar = "^3.8.1"
psycopg2-binary = "^2.9.5"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
lorem = "^0.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
